[Refactor] #11 - Detail Weather List 화면 VC, View 분할
